// src/app/models/advert.model.ts
export class advert {
    id?: string;
    userId?: string;
    userDescription?: string;
    dateExpiry?: Date;
    advertPackageId?: string;
    advertPackageDescription?: string;
    title?: string;
    subTitle?: string;
    link?: string;
    content?: string;
    statusId?: string;
    statusDescription?: string;
    avatar?: string;
}